lucy/requirements.txt
flask==2.0.1
gunicorn==20.1.0
google-cloud-secret-manager==2.12.0
openai==1.0.0